LIBFCREPO-1288. Added "presentation_set_label" field as "docValue"
Added "presentation_set_label" field as a "docValue" to the
"Facets Fields" list. As a "docValue", the "sorted" and "indexed" fields
can be set to "false".

The "useDocValuesAsStored" is set to "true" because that is the default
in Solr schema 1.6 (we are currently using Solr schema v1.5), so that
the field value will be returned in Solr "fl" wildcard queries.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1288